修正时间为int32。

+++ das/src/core/entity/wljson.go
@@ -117,7 +117,7 @@ type AddDevUserStep struct {
 	SubOpen   int    `json:"subOpen"`   // 次开锁方式 (0-正常指纹，1-胁迫指纹, 0:正常密码，1:胁迫密码，2:时间段密码，3:远程密码）
 	Step      int    `json:"step"`      // 步骤序号（指纹：需要4步，1，2，3，4分别代表上下左右；刷卡：需要1步；密码：需要2步，分别是第一次输入密码和第二次输入密码）
 	StepState int    `json:"stepState"` // 0表示成功，1表示失败
-	Time      int64  `json:"time"`
+	Time      int32  `json:"time"`
 }
 
 //4. 删除设备用户（APP-->后台-->锁）
@@ -133,7 +133,7 @@ type DelDevUser struct {
 	UserId   uint16 `json:"userId"`   // 设备用户ID
 	MainOpen uint8  `json:"mainOpen"` // 主开锁方式（1-密码，2-刷卡，3-指纹）
 	SubOpen  uint8  `json:"subOpen"`  // 次开锁方式 (0-正常指纹，1-胁迫指纹, 0:正常密码，1:胁迫密码，2:时间段密码，3:远程密码）
-	Time     int64  `json:"time"`
+	Time     int32  `json:"time"`
 	Bindid   string `json:"bindid,omitempty"`  // zigbee锁网关账号
 	Bindstr  string `json:"bindstr,omitempty"` // zigbee锁网关密码
 }
@@ -378,7 +378,7 @@ type OpenLockLog struct {
 	SubOpen   uint8  `json:"subOpen"`   // 次开锁方式 (0-正常指纹，1-胁迫指纹, 0:正常密码，1:胁迫密码，2:时间段密码，3:远程密码）
 	SinMul    uint8  `json:"sin_mul"`   // 开门模式（1：表示单人模式, 2：表示双人模式）
 	Remainder uint16 `json:"remainder"` // 临时用户剩下能开门的次数,其他用户为0xffff
-	Time      int64  `json:"time"`
+	Time      int32  `json:"time"`
 }
 type UploadOpenLockLog struct {
 	Cmd     int    `json:"cmd"`
@@ -400,7 +400,7 @@ type EnterMenu struct {
 	MainOpen uint8  `json:"mainOpen"` // 主开锁方式（1-密码，2-刷卡，3-指纹）
 	SubOpen  uint8  `json:"subOpen"`  // 次开锁方式 (0-正常指纹，1-胁迫指纹, 0:正常密码，1:胁迫密码，2:时间段密码，3:远程密码）
 	SinMul   uint8  `json:"sin_mul"`  // 开门模式（1：表示单人模式, 2：表示双人模式）
-	Time     int64  `json:"time"`
+	Time     int32  `json:"time"`
 }
 type UploadEnterMenuLog struct {
 	Cmd     int    `json:"cmd"`
@@ -424,7 +424,7 @@ type AlarmMsg struct {
 	Vendor  string `json:"vendor"`
 	SeqId   int    `json:"seqId"`
 
-	Time int64 `json:"time"`
+	Time int32 `json:"time"`
 }
 
 // 低电压告警
@@ -437,7 +437,7 @@ type AlarmMsgBatt struct {
 	SeqId   int    `json:"seqId"`
 
 	Value int   `json:"value"` // 电量百分比 低压报警带有电池电压告警
-	Time  int64 `json:"time"`
+	Time  int32 `json:"time"`
 }
 
 //14. 锁激活状态上报
@@ -503,7 +503,7 @@ type DoorBellCall struct {
 	Vendor  string `json:"vendor"`
 	SeqId   int    `json:"seqId"`
 
-	Time int64 `json:"time"`
+	Time int32 `json:"time"`
 
 	Bindid  string `json:"bindid"`  // zigbee锁网关账号
 	Bindstr string `json:"bindstr"` // zigbee锁网关密码
